Add Factory Classes for Task Creation & rename Task Classes
The code currently mixes task creation logic within the task classes
(Todos, Deadlines, Events), making it harder to maintain and extend
as more task types are introduced. Additionally, the plural class
names are misleading since each class represents a single task.

To improve modularity and clarity, factory classes (TodoFactory,
DeadlineFactory, EventFactory) are introduced to handle task creation
logic separately from the task functionality. Task classes are
renamed to Todo, Deadline, and Event to better reflect that each
represents a single instance.

This refactor simplifies the code by decoupling task creation
from task behaviour, making it easier to manage, extend, and test.
It also improves clarity and consistency by using singular naming
conventions for task classes.
