+++ src/main/java/alfred/task/Deadline.java
@@ -4,8 +4,6 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import alfred.exception.AlfredException;
 
@@ -13,8 +11,7 @@ import alfred.exception.AlfredException;
  * Represents a Deadline task in the task management system.
  * A Deadline task has a description and a deadline date.
  */
-public class Deadlines extends Task {
-    private static final String VALID_DEADLINE_FORMAT = "^deadline\\s+(.+?)\\s+/by\\s+(\\d{4}-\\d{2}-\\d{2})$";
+public class Deadline extends Task {
     private static final DateTimeFormatter DEADLINE_DATE_FORMAT = DateTimeFormatter.ofPattern("MMM d yyyy");
     private LocalDate deadline;
 
@@ -25,7 +22,7 @@ public class Deadlines extends Task {
      * @param deadline    The deadline date in yyyy-MM-dd format.
      * @throws AlfredException If the deadline date is not in the correct format.
      */
-    public Deadlines(String description, String deadline) throws AlfredException {
+    public Deadline(String description, String deadline) throws AlfredException {
         super(description);
         this.deadline = parseDeadline(deadline);
     }
@@ -39,7 +36,7 @@ public class Deadlines extends Task {
      * @param tags        A list of tags associated with the task.
      * @throws AlfredException If the deadline date is not in the correct format.
      */
-    public Deadlines(String description, String deadline, boolean isDone, List<String> tags) throws AlfredException {
+    public Deadline(String description, String deadline, boolean isDone, List<String> tags) throws AlfredException {
         super(description, isDone, tags);
         this.deadline = parseDeadline(deadline);
     }
@@ -76,62 +73,6 @@ public class Deadlines extends Task {
      */
     private String formatDeadlineForDisplay() {
         return deadline.format(DEADLINE_DATE_FORMAT);
-    }
-
-    /**
-     * Creates a Deadline task based on the provided task data and completion status.
-     * It verifies that the input data matches the expected format for a Deadline task.
-     *
-     * @param taskData A String array containing the task details.
-     * @param isDone A boolean indicating whether the task is completed.
-     * @param tags A list that contains tags task is associated with.
-     * @return A <code>Deadlines</code> object created from the provided task data.
-     * @throws AlfredException If the task data is corrupted or does not match the expected Deadline format.
-     */
-    public static Deadlines createTaskFromData(String[] taskData, boolean isDone,
-                                               List<String> tags) throws AlfredException {
-        if (taskData.length != 5) {
-            throw new AlfredException("Corrupted save: Invalid deadline format");
-        }
-        String description = taskData[3];
-        String deadline = taskData[4];
-        return new Deadlines(description, deadline, isDone, tags);
-    }
-
-    /**
-     * Creates a Deadlines task from the input string.
-     * The input should be in the format: deadline task /by yyyy-mm-dd.
-     *
-     * @param input The input string containing the task details.
-     * @return A Deadlines object with the specified description and deadline date.
-     * @throws AlfredException If the input string does not match the expected format.
-     */
-    public static Task createTaskFromInput(String input) throws AlfredException {
-        String[] parsedInput = parseInputForDeadline(input);
-        String description = parsedInput[0];
-        String deadline = parsedInput[1];
-
-        return new Deadlines(description, deadline);
-    }
-
-    /**
-     * Parses the input string to extract the task description and deadline.
-     * Validates the input format using a regular expression.
-     *
-     * @param input The input string to be parsed.
-     * @return An array where the first element is the description and the second element is the deadline.
-     * @throws AlfredException If the input format is incorrect.
-     */
-    private static String[] parseInputForDeadline(String input) throws AlfredException {
-        Pattern pattern = Pattern.compile(VALID_DEADLINE_FORMAT);
-        Matcher matcher = pattern.matcher(input);
-
-        if (matcher.matches()) {
-            return new String[]{matcher.group(1), matcher.group(2)};
-        } else {
-            throw new AlfredException("That is the wrong deadline format Sir. It goes deadline <task> "
-                    + "/by yyyy-mm-dd");
-        }
     }
 
     /**
